Add (optional) support for loading modules
Add support for loading modules in tiny-initramfs. The support is
extremely simple, it just looks for a file /modules in the initramfs
image, where each line should be the absolute path (within the
initramfs image) to the module file name, followed optionally by a
space and the options.

No dependency resolution is performed, the modules are loaded in the
order they are specified in the /modules file. The construction of the
initramfs image should take care of the proper order of the modules and
adding all required dependencies.

If a module cannot be loaded, an error message will be shown, but
tiny-initramfs will still try to mount the root file system.